+++ packages/generator/src/openapi/fileGenerator.ts
@@ -2,9 +2,11 @@ import { Project } from "ts-morph";
 import { createProgressBar } from "../progressBar";
 import Entity from "./models/entity";
 import Enum from "./models/enum";
+import ObjectEntity from "./models/objectEntity";
 import { IGeneratorParams } from "./types";
-import EntityWriter from "./writers/entityWriter";
 import EnumWriter from "./writers/enumWriter";
+import ObjectEntityWriter from "./writers/objectEntityWriter";
+import TypeEntityWriter from "./writers/typeEntityWriter";
 
 export default class FileGenerator {
   constructor(private project: Project, private entities: Entity[], private enums: Enum[]) {}
@@ -23,9 +25,15 @@ export default class FileGenerator {
     await this.project.save();
     progress.increment();
 
-    const entityWriter = new EntityWriter(directory, params);
+    const objectEntityWriter = new ObjectEntityWriter(directory, params);
+    const typeEntityWriter = new TypeEntityWriter(directory);
     for (const definition of this.entities) {
-      entityWriter.write(definition);
+      if (definition instanceof ObjectEntity) {
+        objectEntityWriter.write(definition);
+      } else {
+        typeEntityWriter.write(definition);
+      }
+
       progress.increment();
     }
     await this.project.save();