+++ packages/generator/src/openapi/models/typeEntity.ts
@@ -1,0 +1,5 @@
+import TypeDefinition from "./typeDefinition";
+
+export default class TypeEntity {
+  constructor(public name: string, public type: TypeDefinition) {}
+}